+++ src/components/atoms/Button/style.ts
@@ -1,0 +1,10 @@
+import Styled from 'styled-components';
+
+export const SButton = Styled.button`
+  border: 0;
+  background-color: inherit;
+  padding : 0;
+  :focus {
+    outline : none;
+  }
+`;
