+++ src/components/atoms/Img/index.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import Styled from 'styled-components';
-import { palette } from 'styled-tools';
+import { SDiv, SImg } from './style';
 
 export interface IProps {
   alt?: string;
@@ -15,20 +14,5 @@ function Img({ alt = '', src, ...props }: IProps): React.ReactElement {
 
   return <SImg src={src} alt={alt} {...props} />;
 }
-
-const SImg = Styled.img`
-  width: inherit;
-  height: inherit;
-  object-fit: inherit;
-  border-radius: inherit;
-`;
-
-const SDiv = Styled.div`
-  width: inherit;
-  height: inherit;
-  object-fit: inherit;
-  border-radius: inherit;
-  background-color: ${palette('emptyImage')};
-`;
 
 export default Img;
