+++ src/components/atoms/Label/style.ts
@@ -1,0 +1,5 @@
+import Styled from 'styled-components';
+
+export const SLabel = Styled.label`
+  margin : 0;
+`;
