+++ src/components/molecules/MyPageSection/index.tsx
@@ -1,10 +1,9 @@
 import { MoreViewArrow } from 'assets/images';
 import { Icon, Link } from 'components/atoms';
 import React from 'react';
-import Styled from 'styled-components';
-import { palette, prop } from 'styled-tools';
 import { IConcert } from 'types/concert.type';
 import { IChallenge } from 'types/myPage.type';
+import { EngTitle, Title, Wrapper } from './style';
 
 export interface IProps {
   title: string;
@@ -19,13 +18,13 @@ export interface IProps {
 function MyPageSection({ title, subTitle, path, data, renderItemList, ...props }: IProps): React.ReactElement {
   return (
     <Wrapper {...props}>
-      <h2 className={title === '작성한 글' ? 'dp2' : 'dp2_eng'}>{title}</h2>
+      {title === '작성한 글' ? <Title>{title}</Title> : <EngTitle>{title}</EngTitle>}
       <div className="infoContainer">
         <h2 className="body3">{subTitle}</h2>
         <Link to={path}>
           <div className="infoContainer__link">
             <p className="body2">더보기</p>
-            <Icon src={MoreViewArrow} />
+            <Icon src={MoreViewArrowIcon} />
           </div>
         </Link>
       </div>
@@ -39,53 +38,5 @@ function MyPageSection({ title, subTitle, path, data, renderItemList, ...props }
     </Wrapper>
   );
 }
-
-const Wrapper = Styled.div<{ column: number; gap: number }>`
-  width: 1062px;
-
-  .dp2, .dp2_eng {
-    margin-bottom: 14px;
-    color: ${palette('grayscale', -3)};
-  }
-
-  .infoContainer {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    margin-bottom: 60px;
-    color: ${palette('grayscale', -4)};
-
-    &__link {
-      display: flex;
-      align-items: center;
-      width: 60px;
-      height: 25px;
-      color: ${palette('grayscale', 4)};
-
-      p {
-        margin: 2px 0 1px 0;
-      }
-
-      img {
-        width: 24px;
-        height: 24px;
-      }
-    }
-  }
-  
-  .emptyContainer {
-    width: 100%;
-    text-align: center;
-    margin-top: 90px;
-    color: ${palette('grayscale', 2)}
-  }
-
-  .itemContainer {
-    display: grid;
-    grid-template-columns: repeat(${prop('column')}, 1fr);
-    gap: ${prop('gap')}px;
-    width: 100%;
-  }
-`;
 
 export default MyPageSection;
