+++ src/components/organisms/MyPageHeader/index.tsx
@@ -4,9 +4,8 @@ import { MyPageSlider } from 'components/molecules';
 import React from 'react';
 import { useRecoilValue } from 'recoil';
 import { userState } from 'stores/user';
-import Styled from 'styled-components';
-import { palette } from 'styled-tools';
 import { IMyPageHeader } from 'types/myPage.type';
+import { Wrapper } from './style';
 
 export interface IProps {
   userInfo: IMyPageHeader | null;
@@ -41,42 +40,5 @@ function MyPageHeader({ userInfo }: IProps): React.ReactElement {
     </Wrapper>
   );
 }
-
-const Wrapper = Styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  .dp1 {
-    color: ${palette('grayscale', 0)};
-  }
-
-  .dp1_eng {
-    -webkit-background-clip: text;
-    -webkit-text-fill-color: transparent;
-    background-clip: text;
-    background-image: linear-gradient(to right, ${palette('primary', 3)}, ${palette('primary', 0)});
-  }
-
-  .userInfo {
-    margin-right: 50px;
-
-    &__thumbnail {
-      width : 74px;
-      height : 74px;
-      margin-bottom: 36px;
-      border: solid 1px ${palette('grayscale', 0)};
-    }
-
-    &__editIcon {
-      margin: 10px 0 0 20px;
-    }
-
-    &__tag {
-      margin-top: 40px;
-      color: ${palette('grayscale', 1)};
-    }
-  }
-`;
 
 export default MyPageHeader;
